Extract impact summary from preflight and report to separate component
