+++ src/features/import/components/ImportDialog/Preflight.tsx
@@ -1,16 +1,7 @@
 import { FC } from 'react';
-import {
-  Box,
-  CircularProgress,
-  Stack,
-  Typography,
-  useTheme,
-} from '@mui/material';
+import { Box, CircularProgress, Typography, useTheme } from '@mui/material';
 
-import AddedOrgs from './elements/AddedOrgs';
-import AddedTags from './elements/AddedTags';
-import ChangedFields from './elements/ChangedFields';
-import CreatedAndUpdated from './elements/CreatedAndUpdated';
+import ImpactSummary from './elements/ImpactSummary';
 import ImportFooter from './elements/ImportFooter';
 import ImportHeader from './elements/ImportHeader';
 import ImportMessageList from './elements/ImportMessageList';
@@ -44,8 +35,6 @@ const Preflight: FC<PreflightProps> = ({
   }
 
   const {
-    orgsWithNewPeople,
-    addedTags,
     importDisabled,
     importPeople,
     loading,
@@ -93,26 +82,7 @@ const Preflight: FC<PreflightProps> = ({
               <Typography sx={{ mb: 2 }} variant="h5">
                 <Msg id={messageIds.preflight.headers.summary} />
               </Typography>
-              <CreatedAndUpdated summary={summary} />
-              <Stack spacing={2} sx={{ mt: 2 }}>
-                <ChangedFields
-                  changedFields={summary.updated.byChangedField}
-                  initializedFields={summary.updated.byInitializedField}
-                  orgId={orgId}
-                />
-                {addedTags.length > 0 && (
-                  <AddedTags
-                    addedTags={addedTags}
-                    numPeopleWithTagsAdded={summary.tagged.total}
-                  />
-                )}
-                {orgsWithNewPeople.length > 0 && (
-                  <AddedOrgs
-                    numPeopleWithOrgsAdded={summary.addedToOrg.total}
-                    orgsWithNewPeople={orgsWithNewPeople}
-                  />
-                )}
-              </Stack>
+              <ImpactSummary orgId={orgId} summary={summary} />
             </Box>
             <Box ml={2} sx={{ overflowY: 'auto' }} width="50%">
               <Typography sx={{ mb: 2 }} variant="h5">
